refactor: modularize text extraction utilities
